refactor: Modularize JIRA MCP implementation - Split into models, core, and operations modules with proper validation
