+++ src/models/__init__.py
@@ -1,0 +1,21 @@
+"""
+JIRA MCP model definitions.
+"""
+from .comment import CommentArgs, GetCommentsArgs
+from .worklog import LogWorkArgs
+from .issue import (
+    IssueType,
+    IssueArgs,
+    IssueTransitionArgs,
+    CloneIssueArgs,
+)
+
+__all__ = [
+    'CommentArgs',
+    'GetCommentsArgs',
+    'LogWorkArgs',
+    'IssueType',
+    'IssueArgs',
+    'IssueTransitionArgs',
+    'CloneIssueArgs',
+]
