+++ src/operations/__init__.py
@@ -1,0 +1,28 @@
+"""
+JIRA operations implementation.
+"""
+from .issues import (
+    get_issue,
+    search_issues,
+    create_issue,
+    update_issue,
+    clone_issue,
+)
+from .comments import (
+    add_comment,
+    get_comments,
+)
+from .worklog import log_work
+from .projects import get_projects
+
+__all__ = [
+    'get_issue',
+    'search_issues',
+    'create_issue',
+    'update_issue',
+    'clone_issue',
+    'add_comment',
+    'get_comments',
+    'log_work',
+    'get_projects',
+]
